Updated build script to not require specific path position

+++ buildrpm.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 set -e
 spec=$1
-[ -f "$spec" ] || echo "Spec $1 not found"
-cd /root
-rm -rf anaconda-* original-*
+if [ -f "$spec" ]; then
+	echo "Spec $1 not found"
+	exit 1
+fi
 dnf -y upgrade
 dnf -y install dnf-plugins-core rpmdevtools
 dnf -y copr enable kwetlesen/libmdbx
+cd /root
 rpmdev-setuptree
+cd -
 dnf -y builddep $spec
 if [ -f builds.txt ]; then
 	echo Run these commands:
